working on initialisation and overworld sprites
this is a fairly small commit because im going to try to commit at least something every day. even it if comes to only a sprite change or something or a 'working on' commit like this one. I just want something to be pushed every day for the next month for this project.