created camera object
i also expanded the world to make it larger to test the camera. screen shake is also completed, but not being used right now due to a player speed bug and theres nothing to cause the screen to shake at the moment. i also created a icon, title, and description for the game, its a work in progress, and not final.